refactor js filter methods to helpers
